refactor(rust): consolidate OCR in core
