+++ litellm-rust/crates/ai-gateway/Cargo.toml
@@ -23,7 +23,6 @@ tokio = { workspace = true, features = ["rt-multi-thread", "macros", "net", "tim
 tokio-tungstenite.workspace = true
 futures-util.workspace = true
 serde_json.workspace = true
-base64.workspace = true
 axum = { workspace = true, features = ["ws"], optional = true }
 serde.workspace = true
 subtle = { workspace = true, optional = true }
