+++ litellm-rust/crates/ai-gateway/src/integrations/mod.rs
@@ -1,12 +1,5 @@
-//! Pure-Rust logging integrations. Names map 1:1 to Python
-//! `litellm/integrations/`:
-//!   - [`custom_guardrail::CustomGuardrail`] — the guardrail callback trait
-//!   - [`custom_logger::CustomLogger`]  — the callback trait
-//!   - [`litellm_python_proxy_api::LiteLLMPythonProxyAPILogger`] — ships events
-//!     to the Python proxy's `/v1/rust_control_plane/logs` endpoint
-//!   - [`types`] — the typed `StandardLoggingPayload` wire contract
+//! Host-side integrations for the Rust gateway.
+//!
+//! Callback and guardrail contracts live in `litellm_core::callbacks`.
 
-pub mod custom_guardrail;
-pub mod custom_logger;
 pub mod litellm_python_proxy_api;
-pub mod types;
