+++ litellm-rust/crates/ai-gateway/src/io/mod.rs
@@ -1,5 +1,4 @@
 pub mod audio_transcription;
-pub mod ocr;
 pub mod realtime;
 pub mod realtime_pool;
 pub mod responses_ws;
