--- litellm-rust/crates/ai-gateway/src/io/ocr.rs
@@ -1,1 +1,0 @@
-pub use crate::ocr::{OcrRequest, ocr};
