+++ litellm-rust/crates/ai-gateway/src/realtime/streaming.rs
@@ -13,10 +13,10 @@ use litellm_core::realtime::types::RealtimeEvent;
 use serde_json::Value;
 
 use crate::constants::DEFAULT_PROVIDER;
-use crate::integrations::custom_logger::{
+use litellm_core::callbacks::custom_logger::{
     CallbackTiming, CallbackValue, CustomLogger, CustomLoggerRunner, LoggingError, ModelCallDetails,
 };
-use crate::integrations::types::{
+use litellm_core::callbacks::types::{
     RequestMetadata, StandardLoggingMetadata, StandardLoggingPayload, Usage,
 };
 
@@ -232,8 +232,8 @@ impl RealTimeStreaming {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::integrations::custom_logger::LogError;
-    use crate::integrations::custom_logger::LogFuture;
+    use litellm_core::callbacks::custom_logger::LogError;
+    use litellm_core::callbacks::custom_logger::LogFuture;
     use std::sync::atomic::{AtomicU64, Ordering};
 
     fn event(raw: &str) -> RealtimeEvent {
