+++ litellm-rust/crates/core/src/callbacks/types.rs
