+++ litellm-rust/crates/core/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod audio_transcription;
 pub mod caching;
 pub mod call_lifecycle;
+pub mod callbacks;
 pub mod chat_completions;
 pub mod constants;
 pub mod error;
