+++ litellm-rust/crates/core/src/ocr/common_utils.rs
@@ -1,24 +1,24 @@
 use std::net::IpAddr;
 use std::time::{Duration, Instant};
 
+use crate::CoreResult;
+use crate::error::CoreError;
+use crate::ocr::transformation::OcrProviderConfig;
 use base64::Engine;
 use base64::engine::general_purpose::STANDARD as BASE64_STANDARD;
-use litellm_core::CoreResult;
-use litellm_core::error::CoreError;
-use litellm_core::ocr::transformation::OcrProviderConfig;
 use reqwest::Url;
 use serde_json::{Map, Value};
 
-use litellm_core::providers::azure_ai::ocr::transformation::{
+use crate::providers::azure_ai::ocr::transformation::{
     AZURE_AI_OCR_CONFIG, AZURE_DOCUMENT_INTELLIGENCE_OCR_CONFIG,
 };
-use litellm_core::providers::mistral::ocr::transformation::MISTRAL_OCR_CONFIG;
-use litellm_core::providers::vertex_ai::ocr::transformation as vertex_ai;
-use litellm_core::providers::vertex_ai::ocr::transformation::{
+use crate::providers::mistral::ocr::transformation::MISTRAL_OCR_CONFIG;
+use crate::providers::vertex_ai::ocr::transformation as vertex_ai;
+use crate::providers::vertex_ai::ocr::transformation::{
     VERTEX_AI_DEEPSEEK_OCR_CONFIG, VERTEX_AI_OCR_CONFIG,
 };
 
-use crate::client::http_client;
+use super::client::http_client;
 
 const ERROR_BODY_MAX_CHARS: usize = 256;
 const AZURE_DOCUMENT_INTELLIGENCE_POLL_TIMEOUT_SECS: u64 = 120;
@@ -67,7 +67,7 @@ pub(super) fn string_headers(
                 .ok_or_else(|| {
                     CoreError::InvalidRequest(format!(
                         "OCR extra_headers.{key} must be a string, got {}",
-                        litellm_core::error::json_type_name(&value)
+                        crate::error::json_type_name(&value)
                     ))
                 })
         })
